Create a Traceability Matrix

Returns a table that links all exported functions and their aliases to their documentation (`man` files), the R scripts containing them, and the test scripts that reference them.

Usage

create_traceability_matrix(
  pkg_name,
  pkg_source_path,
  func_covr = NULL,
  verbose = FALSE,
  execute_coverage = TRUE
)

Arguments

pkg_name

name of package

pkg_source_path

path to a source package

func_covr

function coverage

verbose

Logical (`TRUE`/`FALSE`). If `TRUE`, show any warnings/messages per function.

execute_coverage

Logical (`TRUE`/`FALSE`). If `TRUE`, execute test coverage.

Value

a nested list with fine grained traceability matrices

Examples

if (FALSE) { # \dontrun{
# set CRAN repo to enable running of reverse dependencies
r = getOption("repos")
r["CRAN"] = "http://cran.us.r-project.org"
old <- options(repos = r)

dp <- system.file("test-data", "here-1.0.1.tar.gz", 
   package = "risk.assessr")

# Set up the package using the temporary file
install_list <- set_up_pkg(dp)

# Extract information from the installation list
build_vignettes <- install_list$build_vignettes
package_installed <- install_list$package_installed
pkg_source_path <- install_list$pkg_source_path
rcmdcheck_args <- install_list$rcmdcheck_args

# check if the package needs to be installed locally
package_installed <- install_package_local(pkg_source_path)

# Get package name and version
pkg_desc <- get_pkg_desc(pkg_source_path, 
                         fields = c("Package", 
                                    "Version"))
pkg_name <- pkg_desc$Package
package_installed <- TRUE
covr_list <- test.assessr::get_package_coverage(
  pkg_source_path,
  package_installed
)
# run without test coverage   
tm_list_no_covr <- create_traceability_matrix(pkg_name, 
                                      pkg_source_path,
                                      execute_coverage = FALSE)   
                                      
# run with test coverage                                         
tm_list_covr <- create_traceability_matrix(pkg_name, 
                                      pkg_source_path,
                                      covr_list$res_cov)
                                      
# run with test coverage parameter but no test coverage list                                        
tm_list_no_covr_list <- create_traceability_matrix(pkg_name, 
                                      pkg_source_path)   
                                         
options(old)
} # }